Update scrollscript.js
changed the scroll nav to enable at the bottom of the page, looked into adding the fade out when the hide-nav is removed. It's not as simple as having a secondary class with the opacity fade in and adding it to the else statement while removing it from the if.. I'll figure it out soon.