Update styles.css
Really got a lot done in the styling, adjusted the placing of the stussy S to fit in with the view... still really jumpy when not at 100% viewport. Customized the nav buttons a little, they didn't fit well before. Got the hero-bio dialed in with all viewports. Deleted a load of things I was no longer using.